Highlighter
# highlighter
This Chrome extension increases productivity by finding search results quicker. Just click on the extension icon before clicking on a link in the Google search results. The extension will grab the description text under the link, and automatically highlight the text on newly opened tab and scroll down to it.